+++ ColorPicker/SwatchButton.js
@@ -8,7 +8,6 @@
 import kind from '@enact/core/kind';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
-import React from 'react';
 
 import {ButtonBase, ButtonDecorator} from '../Button';
 import Skinnable from '../Skinnable';