+++ Input/InputDecoratorIcon.js
@@ -1,7 +1,5 @@
-
 import kind from '@enact/core/kind';
 import PropTypes from 'prop-types';
-import React from 'react';
 import onlyUpdateForKeys from 'recompose/onlyUpdateForKeys';
 
 import Icon from '../Icon';