+++ Switch/tests/Switch-specs.js
@@ -1,7 +1,6 @@
-import {mount} from 'enzyme';
-import React from 'react';
+import {mount, shallow} from 'enzyme';
 
-import Switch from '../Switch';
+import Switch, {SwitchBase} from '../Switch';
 
 describe('Switch Specs', () => {
 	test('should have `selected` if selected prop is true', () => {
@@ -15,26 +14,24 @@ describe('Switch Specs', () => {
 		expect(actual).toBe(expected);
 	});
 
-	// At this moment, Switch component is based on internal/ToggleIcon. After
-	// Sandstone architecture sync-up, below tests need to be modified
 	test('should have animated class by default', () => {
-		const subject = mount(
-			<Switch />
+		const subject = shallow(
+			<SwitchBase />
 		);
 
 		const expected = true;
-		const actual = subject.find('ToggleIcon').hasClass('animated');
+		const actual = subject.hasClass('animated');
 
 		expect(actual).toBe(expected);
 	});
 
 	test('should not have animated class when noAnimation prop is true', () => {
-		const subject = mount(
-			<Switch noAnimation />
+		const subject = shallow(
+			<SwitchBase noAnimation />
 		);
 
 		const expected = false;
-		const actual = subject.find('ToggleIcon').hasClass('animated');
+		const actual = subject.hasClass('animated');
 
 		expect(actual).toBe(expected);
 	});