+++ samples/sampler/stories/default/LabeledIconButton.js
@@ -1,7 +1,6 @@
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {boolean, select, text, number} from '@enact/storybook-utils/addons/knobs';
 import {LabeledIconBase as UiLabeledIconBase, LabeledIcon as UiLabeledIcon} from '@enact/ui/LabeledIcon';
-import React from 'react';
 import {storiesOf} from '@storybook/react';
 
 import LabeledIconButton from '@enact/agate/LabeledIconButton';
@@ -25,7 +24,6 @@ storiesOf('Agate', module)
 				selected={boolean('selected', Config)}
 				size={select('size', ['smallest', 'small', 'large', 'huge'], Config)}
 				spriteCount={number('spriteCount', Config, {min: 1}, 1)}
-				tooltipText={text('tooltipText', Config, 'This is a Labeled Icon Button')}
 			>
 				{text('children', Config, 'Hello LabeledIconButton')}
 			</LabeledIconButton>